Update guessgame
Hi I am new to programming too! I just came across your page. One of my exercises was to view other people's code and see if I can contribute or write it in another way. I definitely like your guess game script. I have written it in a more simple way here. Like I said I'm a noob so maybe the way you wrote it was to get into a good habit or have a better understanding I can't be to sure. But here is my version which is basically the same :). Also I'm not trying to edit yours, you can deny this proposal since I'm not really trying to make you change it. I'm a noob at this website and I wasn't sure how to just comment on the code.